+++ pkg/certmanageroperator/yaml/cert-manager-operator.yaml
